Fix hotkey conflict handling and silent failure paths (closes #7)
- Hotkey conflicts no longer drop all working hotkeys — each profile now
  registers independently, so only the conflicting one is skipped
- Each conflict now shows its own notification instead of only the first
- Silent empty catch on RegisterProfile replaced with proper error logging
- Hotkeys with invalid VirtualKeyCode values (from hand-edited config) are
  now skipped safely instead of being passed to the Windows API
- Modifier flag constants now reference WinApi.MOD_* instead of inline hex
- GlobalAddAtom P/Invoke corrected to return ushort (the actual Win32 ATOM type)
- Removed ToDisplayString blanket catch (method never throws)
- Removed Refresh() alias that was identical to RegisterAll()

+++ VibeSwitcher/NativeMethods/WinApi.cs
@@ -20,7 +20,7 @@ internal static class WinApi
     public static extern bool UnregisterHotKey(IntPtr hWnd, int id);
 
     [DllImport("kernel32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
-    public static extern int GlobalAddAtom(string lpString);
+    public static extern ushort GlobalAddAtom(string lpString);
 
     [DllImport("kernel32.dll", SetLastError = true)]
     public static extern ushort GlobalDeleteAtom(int nAtom);
